# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/register/email": {
      "post": {
        "operationId": "AuthController_registerByEmail",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterByEmailDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/register/phone": {
      "post": {
        "operationId": "AuthController_registerByPhone",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterByPhoneDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/login/email": {
      "post": {
        "operationId": "AuthController_loginByEmail",
        "summary": "Логин по email + пароль",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginByEmailDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/login/phone": {
      "post": {
        "operationId": "AuthController_loginByPhone",
        "summary": "Логин по телефону + пароль",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginByPhoneDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/login/email/otp/request": {
      "post": {
        "operationId": "AuthController_requestEmailOtp",
        "summary": "Запросить OTP-код для email",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestEmailOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "ok": true
                  }
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/login/phone/otp/request": {
      "post": {
        "operationId": "AuthController_requestPhoneOtp",
        "summary": "Запросить OTP-код для телефона",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestPhoneOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "ok": true
                  }
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/login/phone/otp/verify": {
      "post": {
        "operationId": "AuthController_verifyPhoneOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyPhoneOtpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/login/email/otp/verify": {
      "post": {
        "operationId": "AuthController_verifyEmailOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailOtpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "summary": "Обновить access-токен по refresh (cookie)",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ],
        "security": [
          {
            "cookie": []
          }
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "summary": "Выйти и отозвать сессию",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "ok": true
                  }
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ],
        "security": [
          {
            "cookie": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "AI Platform API",
    "description": "Prompts / Generations / Users",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "UserRole": {
        "type": "string",
        "enum": [
          "user",
          "admin"
        ]
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "email": {
            "type": "string",
            "example": "user@example.com",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "example": "+79998887766",
            "nullable": true
          },
          "role": {
            "example": "user",
            "$ref": "#/components/schemas/UserRole"
          },
          "credits": {
            "type": "number",
            "example": 100,
            "description": "Баланс кредитов"
          },
          "isActive": {
            "type": "boolean",
            "example": true,
            "description": "Активен ли пользователь"
          },
          "isEmailVerified": {
            "type": "boolean",
            "example": false
          },
          "isPhoneVerified": {
            "type": "boolean",
            "example": true
          },
          "lastLoginAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "role",
          "credits",
          "isActive",
          "isEmailVerified",
          "isPhoneVerified"
        ]
      },
      "RegisterByEmailDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "example": "StrongPass123!"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RegisterByPhoneDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "+79998887766"
          },
          "password": {
            "type": "string",
            "example": "secretPassword123",
            "description": "Пароль (необязателен, если используется вход по OTP)"
          }
        },
        "required": [
          "phone"
        ]
      },
      "LoginByEmailDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "example": "StrongPass123!"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "user": {
            "$ref": "#/components/schemas/UserDto"
          }
        },
        "required": [
          "accessToken",
          "user"
        ]
      },
      "LoginByPhoneDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "+79998887766"
          },
          "password": {
            "type": "string",
            "example": "Secret123"
          }
        },
        "required": [
          "phone",
          "password"
        ]
      },
      "RequestEmailOtpDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "RequestPhoneOtpDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "+79998887766"
          }
        },
        "required": [
          "phone"
        ]
      },
      "VerifyPhoneOtpDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "+79998887766"
          },
          "code": {
            "type": "string",
            "example": "123456"
          }
        },
        "required": [
          "phone",
          "code"
        ]
      },
      "VerifyEmailOtpDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "code": {
            "type": "string",
            "example": "123456"
          }
        },
        "required": [
          "email",
          "code"
        ]
      },
...
    }
  }
}